samples/python/uff_ssd/requirements.txt
numpy==1.18.1; python_version<"3.8" and platform_system == "Windows"
numpy==1.19.4; python_version<"3.8" and platform_system != "Windows"
numpy==1.20.2; python_version>="3.8"
Pillow; python_version<"3.6"
Pillow==8.3.2; python_version>="3.6"
pycuda<2021.1; python_version<"3.7"
pycuda; python_version>="3.7"
requests
